Fix validation error on birth check
This was causing our tests to fail.

The fault was that we were expecting
date to be a datetime object and
we were trying to cast a str to a datetime
again.

i.e. the birth was already a
datetime object and we were trying to cast
it from str to datetime.